Chapter-18/image_predict_function/requirements.txt
functions-framework==3.*
Pillow
numpy
google-cloud-aiplatform
google-cloud-storage
